better integration of float features to handle position of "new nodes"
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/semagramme/libcaml-grew/trunk@8123 7838e531-6607-4d57-9587-6c381814729c

+++ src/grew_ast.ml
@@ -122,7 +122,7 @@ module Ast = struct
 
   type u_node = {
       node_id: Id.name;
-      position: int option;
+      position: float option;
       fs: feature list;
     }
   type node = u_node * Loc.t
